+++ .changeset/mean-places-float.md
@@ -1,0 +1,5 @@
+---
+'@viamrobotics/svelte-sdk': minor
+---
+
+Add CameraImage and CameraStream components
